+++ Client/Frontend/Home/DownloadsPanel.swift
@@ -41,78 +41,6 @@ struct DownloadedFile: Equatable {
     }
 }
 
-private func getDate(dayOffset: Int) -> Date {
-    let calendar = Calendar(identifier: .gregorian)
-    let components = calendar.dateComponents([.year, .month, .day], from: Date())
-    let today = calendar.date(from: components)!
-    return calendar.date(byAdding: .day, value: dayOffset, to: today)!
-}
-
-struct DateGroupedTableData<T : Equatable> {
-    let todayTimestamp = getDate(dayOffset: 0).timeIntervalSince1970
-    let yesterdayTimestamp = getDate(dayOffset: -1).timeIntervalSince1970
-    let lastWeekTimestamp = getDate(dayOffset: -7).timeIntervalSince1970
-
-    var today: [(T, TimeInterval)] = []
-    var yesterday: [(T, TimeInterval)] = []
-    var lastWeek: [(T, TimeInterval)] = []
-    var older: [(T, TimeInterval)] = []
-
-    var isEmpty: Bool {
-        return today.isEmpty && yesterday.isEmpty && lastWeek.isEmpty && older.isEmpty
-    }
-
-    mutating func add(_ item: T, timestamp: TimeInterval) {
-        if timestamp > todayTimestamp {
-            today.append((item, timestamp))
-        } else if timestamp > yesterdayTimestamp {
-            yesterday.append((item, timestamp))
-        } else if timestamp > lastWeekTimestamp {
-            lastWeek.append((item, timestamp))
-        } else {
-            older.append((item, timestamp))
-        }
-    }
-
-    mutating func remove(_ item: T) {
-        if let index = today.index(where: { item == $0.0 }) {
-            today.remove(at: index)
-        } else if let index = yesterday.index(where: { item == $0.0 }) {
-            yesterday.remove(at: index)
-        } else if let index = lastWeek.index(where: { item == $0.0 }) {
-            lastWeek.remove(at: index)
-        } else if let index = older.index(where: { item == $0.0 }) {
-            older.remove(at: index)
-        }
-    }
-
-    func numberOfItemsForSection(_ section: Int) -> Int {
-        switch section {
-        case 0:
-            return today.count
-        case 1:
-            return yesterday.count
-        case 2:
-            return lastWeek.count
-        default:
-            return older.count
-        }
-    }
-
-    func itemsForSection(_ section: Int) -> [T] {
-        switch section {
-        case 0:
-            return today.map({ $0.0 })
-        case 1:
-            return yesterday.map({ $0.0 })
-        case 2:
-            return lastWeek.map({ $0.0 })
-        default:
-            return older.map({ $0.0 })
-        }
-    }
-}
-
 class DownloadsPanel: UIViewController, UITableViewDelegate, UITableViewDataSource, HomePanel {
     weak var homePanelDelegate: HomePanelDelegate?
     let profile: Profile
